Put hyperparameters stuff in a dedicated class Hyperparametrizable
- SolverDO derive now from Hyperparametrizable so that other class can
share this mechanism.
- LNS solvers main bricks derive also from Hyperparametrizable:
  - ConstraintHandler
  - InitialSolution
  - PostProcessSolution
- We change class names for subsolver related hyperparameters as it can
  be something else than solver from now:
  - SubSolverHyperparameter -> SubBrickHyperparameter
  - SubSolverKwargsHyperparameter -> SubBrickKwargsHyperparameter
